pyproject.toml
[project]
name = "youtubedownloadersummarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.31.0",
    "yt-dlp>=2025.8.20",
]
